Move FilterTranscripts into pipeline_utils and setup v2
